+++ src/types.ts
@@ -459,4 +459,17 @@ export type RNSkiaVideoModule = {
         bitRate: number;
       }[]
     | null;
+
+  /**
+   * Runs the specified function with the JNI class loader.
+   * @platform android
+   */
+  runWithJNIClassLoader?(fn: () => void): void;
+
+  /**
+   * Sleeps for the specified number of microseconds.
+   * @param usec The number of microseconds to sleep.
+   * @platform ios
+   */
+  usleep?(usec: number): void;
 };
